Move ComponentListIterator into its own header.
