+++ OpenSim/Common/ComponentList.h
@@ -9,7 +9,7 @@
  * National Institutes of Health (U54 GM072970, R24 HD065690) and by DARPA    *
  * through the Warrior Web program.                                           *
  *                                                                            *
- * Copyright (c) 2014-2014 Stanford University and the Authors                *
+ * Copyright (c) 2014-2015 Stanford University and the Authors                *
  * Authors: Ayman Habib                                                       *
  * Contributers : Chris Dembia                                                *
  *                                                                            *
@@ -26,8 +26,7 @@
 
 // INCLUDES
 #include <OpenSim/Common/osimCommonDLL.h>
-#include "OpenSim/Common/Object.h"
-#include "Simbody.h"
+#include <Simbody.h>
 
 namespace OpenSim {
 
@@ -62,7 +61,7 @@ public:
 /**
  ComponentFilterMatchAll is a trivial Filter that matches all components.
 */
-class ComponentFilterMatchAll : public ComponentFilter {
+class OSIMCOMMON_API ComponentFilterMatchAll : public ComponentFilter {
 public:
     /** Construct a ComponentFilter that selects all Components. */
     ComponentFilterMatchAll() {}
@@ -88,7 +87,7 @@ public:
  The traversal order is wired at the end of initSystem().
 */
 template <typename T>
-class ComponentList {
+class OSIMCOMMON_API ComponentList {
 public:
     /** A const forward iterator for iterating through ComponentList<T>.
     The const indicates that the iterator provides only 
@@ -139,85 +138,6 @@ private:
     // Internal method to setFilter to ComponentFilterMatchAll if no user specified 
     // filter is provided.
     void setDefaultFilter() { setFilter(ComponentFilterMatchAll()); }
-};
-
-//==============================================================================
-//                            OPENSIM ComponentListIterator
-//==============================================================================
-/** Class used to iterate over subcomponents of a specific type (by default,
- any Component).
- This iterator is const_iterator as it returns only const ref or pointer upon 
- dereferencing.
- This iterator works only in forward direction (not bidirectional)
-
- Use as:
-@code
-ComponentList<GeometryPath> geomPathList = model.getComponentList<GeometryPath>();
-for (const GeometryPath& gpath : geomPathList) {
-    // do something with gpath
-}
-@endcode
-*/
-template <typename T>
-class OSIMCOMMON_API ComponentListIterator :
-    public std::iterator<std::forward_iterator_tag, OpenSim::Component> {
-    friend class ComponentList<T>;
-public:
-    /// Check that the component under the cuurent iterator is same 
-    /// (has same address) as right-hand iter.
-    bool operator==(const ComponentListIterator& iter) const {
-        return _node == &*iter;
-    }
-    /// Check for inequality using same convention as operator==.
-    bool operator!=(const ComponentListIterator& iter) const {
-         return _node != &*iter;
-    }
-    /** Dereference the iterator to get a const ref to Component of proper 
-     type (matching Filter if specified) this is const iterator, use only 
-     const methods. 
-    */
-    const T& operator*() const { return *dynamic_cast<const T*>(_node); } 
-
-    /// Another dereferencing operator that returns a const pointer.
-    const T* operator->() const { return dynamic_cast<const T*>(_node); }
-
-    /** Prefix increment operator to get the next item in the ComponentList.
-     Prefer to use ++iter and not iter++. */
-    ComponentListIterator<T>& operator++();
-
-    /** Postfix increment operator to get the next item in the ComponentList.
-    To enable usage as iter++, although ++iter is more efficient. */
-    ComponentListIterator<T> operator++(int) {
-        ComponentListIterator<T> current = *this;
-        next();
-        return current;
-    }
-
-    /** Method equivalent to pre-increment operator for operator-deficient
-     languages.
-     */
-    ComponentListIterator<T>& next() { return ++(*this); }
-private:
-    // Internal method to advance iterator to next valid component.
-    void advanceToNextValidComponent();
-    // Pointer to current Component that the iterator is processing.
-    const Component* _node;
-    // Root of subtree of Components that we're iterating over.
-    const Component& _root;
-    /** Optional filter to further select Components under _root, defaults to
-    Filter by type. */
-    const ComponentFilter& _filter;
-    /** Constructor that takes a Component and ComponentFilter.
-     The iterator contains a const ref to filter and doesn't take ownership 
-     of it. A pointer is used since iterator at end() doesn't have a valid 
-     Component underneath it.
-     */
-    ComponentListIterator(const Component* node, const ComponentFilter& filter) :
-        _node(node),
-        _root(*node),
-        _filter(filter) {
-        advanceToNextValidComponent(); // in case node is not a match.
-    }
 };
 
 } // end of namespace OpenSim
