+++ OpenSim/Simulation/Test/testFrames.cpp
@@ -33,6 +33,7 @@ Tests Include:
      Add tests here as Frames are added to OpenSim
 
 //=============================================================================*/
+#include <OpenSim/Common/ComponentListIterator.h>
 #include <OpenSim/Simulation/Model/Model.h>
 #include <OpenSim/Simulation/Model/PhysicalOffsetFrame.h>
 #include <OpenSim/Auxiliary/auxiliaryTestFunctions.h>
